requirements.txt
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
hiredis==2.2.3
idna==3.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
rdkit-pypi==2022.9.5
redis==4.5.5
requests==2.30.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
